https://github.com/javers/javers/issues/102 refactor
