+++ javers-persistence-sql/build.gradle
@@ -4,7 +4,7 @@ repositories {
 
 dependencies {
     compile project(':javers-core')
-    compile 'org.polyjdbc:polyjdbc:0.5.0-SNAPSHOT'
+    compile 'org.polyjdbc:polyjdbc:0.5.0'
 
     testCompile project(path: ":javers-core", configuration: "testArtifacts")
     testCompile 'commons-dbcp:commons-dbcp:1.4'
